python37/requirements.txt
aiofile
ckanapi==4.3
functions-framework==1.4.3
gcloud-aio-storage
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas==1.0.4
requests
SQLAlchemy==1.3.13
openpyxl
xlrd